requirements.txt
discord.py
vaderSentiment
aiosqlite
python-dotenv
yt-dlp
PyNaCl
vaderSentiment
google-genai
